1.0_ingesta/requirements.txt
numpy==1.24.3
pandas==1.5.3
openpyxl==3.1.2
kafka-python==2.0.2
avro-python3==1.10.2
python-dateutil==2.8.2
pytz==2023.3 